refactor: fix class file name as per python styling
